Finish part 4

+++ part4/bloglist/controllers/blogs.js
@@ -1,33 +1,52 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const userExtractor = require('../utils/middleware').userExtractor
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({})
+    .populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body
   if (!body.title && !body.url) {
     response.status(400).end()
   } else {
+    const user = request.user
     const blog = new Blog({
       title: body.title,
       author: body.author,
       url: body.url,
-      likes: body.likes || 0
+      likes: body.likes || 0,
+      user: user.id
     })
     const savedBlog = await blog.save()
+    console.log('before:', user.blogs)
+    user.blogs = user.blogs.concat(savedBlog.id)
+    console.log('after:', user.blogs)
+    const savedUser = await user.save()
+    console.log('saved user:', savedUser)
     response.status(201).json(savedBlog)
   }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+  if (user.id !== blog.user.id) {
+    return response.status(401).json({ error: 'wrong user' })
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
 
-blogsRouter.put('/:id', async (request, response) => {
+blogsRouter.put('/:id', userExtractor, async (request, response) => {
+  const user = request.user
+  const existingBlog = await Blog.findById(request.params.id)
+  if (user.id !== existingBlog.user.id) {
+    return response.status(401).json({ error: 'wrong user' })
+  }
   const body = request.body
   const blog = {
     title: body.title,
